fix: pass uuids pt1

+++ src/entities/game_with_finish.rs
@@ -18,7 +18,7 @@ pub struct GameWithFinish {
     pub added_datetime: NaiveDateTime,
     pub updated_datetime: NaiveDateTime,
     pub status: i16,
-    pub rating: i32,
+    pub rating: i16,
     pub notes: String,
     pub finish_date: NaiveDate,
     pub finish_status: i16,
